Move CLI out to separate file
