--- .travis.yml
@@ -1,14 +1,0 @@
-# Config file for automatic testing at travis-ci.org
-
-language: python
-
-python:
-  - "3.3"
-  - "2.7"
-  - "pypy"
-
-# command to install dependencies, e.g. pip install -r requirements.txt --use-mirrors
-install: pip install -U .
-
-# command to run tests, e.g. python setup.py test
-script: py.test
