+++ pyproject.toml
@@ -1,7 +1,3 @@
-[build-system]
-requires = ["hatchling"]
-build-backend = "hatchling.build"
-
 [project]
 name = "caf"
 version = "0.1.1"
@@ -26,6 +22,10 @@ dependencies = [
 ]
 requires-python = ">=3.12"
 
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
+
 [tool.uv]
 dev-dependencies = [
     "pytest",
@@ -38,7 +38,7 @@ dev-dependencies = [
 Homepage = "https://github.com/jamesls/caf"
 
 [project.scripts]
-caf = "caf:main"
+caf = "caf.cli:main"
 
 [tool.hatch.build.targets.wheel]
 packages = ["caf"]
@@ -46,5 +46,40 @@ packages = ["caf"]
 [tool.hatch.version]
 path = "caf/__init__.py"
 
+[tool.ruff]
+line-length = 79
+
+[tool.ruff.format]
+quote-style = "preserve"
+preview = true
+
+[tool.pytest.ini_options]
+addopts = """\
+    --cov caf \
+    --cov-report term-missing \
+    --no-cov-on-fail \
+"""
+testpaths = ["tests"]
+filterwarnings = [
+    "ignore::DeprecationWarning",
+]
+
+[tool.coverage.run]
+branch = true
+
+[tool.coverage.report]
+exclude_lines = [
+    'if TYPE_CHECKING:',
+    'pragma: no cover',
+    'raise NotImplementedError',
+]
+
 [tool.pyright]
-include = ["caf", "tests"]
+include = ["caf"]
+exclude = ["**/__pycache__"]
+
+reportMissingImports = true
+reportMissingTypeStubs = false
+
+pythonVersion = "3.12"
+pythonPlatform = "All"
